# server/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentDisputeRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "AssessmentDisputeRequest",
        "type": "object"
      },
      "AssessmentDisputeResponse": {
        "properties": {
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "goal_id": {
            "title": "Goal Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requested_at": {
            "title": "Requested At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DisputeStatus"
          }
        },
        "required": [
          "id",
          "goal_id",
          "assessment_id",
          "reason",
          "status",
          "requested_at"
        ],
        "title": "AssessmentDisputeResponse",
        "type": "object"
      },
      "AssessmentReevaluateRequest": {
        "additionalProperties": false,
        "properties": {
          "dispute_id": {
            "title": "Dispute Id",
            "type": "string"
          }
        },
        "required": [
          "dispute_id"
        ],
        "title": "AssessmentReevaluateRequest",
        "type": "object"
      },
      "AssessmentResponse": {
        "properties": {
          "ambiguities": {
            "items": {
              "type": "string"
            },
            "title": "Ambiguities",
            "type": "array"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "title": "Assumptions",
            "type": "array"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "title": "Citations",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "cross_question_candidate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cross Question Candidate"
          },
          "derivation_excluded": {
            "title": "Derivation Excluded",
            "type": "boolean"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/AssessmentDimensionResponse"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "disputes": {
            "items": {
              "$ref": "#/components/schemas/AssessmentDisputeDetailResponse"
            },
            "title": "Disputes",
            "type": "array"
          },
          "evaluation_method": {
            "title": "Evaluation Method",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "facts": {
            "items": {
              "type": "string"
            },
            "title": "Facts",
            "type": "array"
          },
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "goal_id": {
            "title": "Goal Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "limitation_labels": {
            "items": {
              "type": "string"
            },
            "title": "Limitation Labels",
            "type": "array"
          },
          "predecessor_assessment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predecessor Assessment Id"
          },
          "provenance_refs": {
            "items": {
              "type": "string"
            },
            "title": "Provenance Refs",
            "type": "array"
          },
          "requested_capability": {
            "title": "Requested Capability",
            "type": "string"
          },
          "revision_invitation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Invitation"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "rubric_id": {
            "title": "Rubric Id",
            "type": "string"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "source_refs": {
            "items": {
              "type": "string"
            },
            "title": "Source Refs",
            "type": "array"
          },
          "state": {
            "$ref": "#/components/schemas/AssessmentState"
          },
          "task_ref": {
            "title": "Task Ref",
            "type": "string"
          },
          "trade_offs": {
            "items": {
              "type": "string"
            },
            "title": "Trade Offs",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "goal_id",
          "evidence_id",
          "run_id",
          "rubric_id",
          "rubric_version",
          "state",
          "task_ref",
          "requested_capability",
          "role",
          "level",
          "evaluation_method",
          "assumptions",
          "source_refs",
          "provenance_refs",
          "facts",
          "trade_offs",
          "citations",
          "ambiguities",
          "feedback",
          "cross_question_candidate",
          "revision_invitation",
          "warnings",
          "limitation_labels",
          "predecessor_assessment_id",
          "derivation_excluded",
          "created_at",
          "dimensions",
          "disputes"
        ],
        "title": "AssessmentResponse",
        "type": "object"
      },
      "AssessmentState": {
        "enum": [
          "feedback-ready",
          "ambiguity-unresolved"
        ],
        "title": "AssessmentState",
        "type": "string"
      },
...
      "ErrorResponse": {
        "description": "The API error envelope.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "current_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current State"
          },
          "field_errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "recovery_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Action"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "correlation_id",
          "retryable"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobRefResponse": {
        "description": "The `202` response for an asynchronous operation.",
        "properties": {
          "deduplicated": {
            "default": false,
            "title": "Deduplicated",
            "type": "boolean"
          },
          "enqueued_at": {
            "title": "Enqueued At",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "kind",
          "status",
          "enqueued_at"
        ],
        "title": "JobRefResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Yuno",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assessments/{assessment_id}": {
      "get": {
        "operationId": "get_assessment_record_api_v1_assessments__assessment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Assessment Record",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/disputes": {
      "post": {
        "operationId": "post_dispute_api_v1_assessments__assessment_id__disputes_post",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentDisputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentDisputeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Post Dispute",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/reevaluate": {
      "post": {
        "operationId": "post_reevaluate_api_v1_assessments__assessment_id__reevaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentReevaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Post Reevaluate",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
...
    "/api/v1/evidence/{evidence_id}": {
      "get": {
        "operationId": "get_evidence_api_v1_evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
    "/api/v1/evidence/{evidence_id}/assess": {
      "post": {
        "operationId": "post_assess_api_v1_evidence__evidence_id__assess_post",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRefResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Post Assess",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
...
    "/api/v1/goals/{goal_id}/evidence": {
      "get": {
        "operationId": "get_goal_evidence_api_v1_goals__goal_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceResponse"
                  },
                  "title": "Response Get Goal Evidence Api V1 Goals  Goal Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Goal Evidence",
        "tags": [
          "evidence-evaluation"
        ]
      },
      "post": {
        "operationId": "post_evidence_api_v1_goals__goal_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Post Evidence",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
    "/api/v1/goals/{goal_id}/learning-state-explanations": {
      "get": {
        "operationId": "get_learning_state_explanations_api_v1_goals__goal_id__learning_state_explanations_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningStateExplanationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Learning State Explanations",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
...
    "/api/v1/goals/{goal_id}/progress": {
      "get": {
        "operationId": "get_goal_progress_api_v1_goals__goal_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Get Goal Progress",
        "tags": [
          "evidence-evaluation"
        ]
      }
    },
...
  }
}